Move jwt auth logic to separate file
